refactoread get msgs

+++ client/inc/client.h
@@ -176,7 +176,7 @@ t_dtp *mx_log_in_request(char *login, char *pass);
 t_dtp *mx_sign_up_request(char *login, char *pass);
 t_dtp *mx_get_rooms_request(long int date);
 t_dtp *mx_log_out_request(char *token);
-t_dtp *mx_get_new_msgs_request(long int date, int room_id);
+t_dtp *mx_get_new_msgs_request(guint64 date, guint64 room_id);
 t_dtp *mx_upd_room_desc_request(int room_id, char *desc); // FOR ROOM UPDATE DESCRIPTION CREATE REQUEST
 t_dtp *mx_upd_room_name_request(int room_id, char *name); // FOR ROOM UPDATE NAME CREATE REQUEST
 t_dtp *mx_upd_user_desc_request(char *desc); // FOR USER DESCRIPTION UPDATE
